Separate presigned url code
